+++ packages/adapter-commons/test/service.test.ts
@@ -1,58 +1,11 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import assert from 'assert';
-import { NotImplemented } from '@feathersjs/errors';
-import { AdapterService, InternalServiceMethods } from '../src';
-import { Params, Id, NullableId } from '@feathersjs/feathers';
+import { MethodService } from './fixture';
 
 const METHODS: [ 'find', 'get', 'create', 'update', 'patch', 'remove' ] = [ 'find', 'get', 'create', 'update', 'patch', 'remove' ];
 
 describe('@feathersjs/adapter-commons/service', () => {
-  class CustomService extends AdapterService {
-  }
-
-  describe('errors when method does not exit', () => {
-    METHODS.forEach(method => {
-      it(`${method}`, () => {
-        const service = new CustomService({});
-
-        // @ts-ignore
-        return service[method]().then(() => {
-          throw new Error('Should never get here');
-        }).catch((error: Error) => {
-          assert.ok(error instanceof NotImplemented);
-          assert.strictEqual(error.message, `Method _${method} not available`);
-        });
-      });
-    });
-  });
-
   describe('works when methods exist', () => {
-    class MethodService extends AdapterService implements InternalServiceMethods {
-      _find (_params?: Params) {
-        return Promise.resolve([]);
-      }
-
-      _get (id: Id, _params?: Params) {
-        return Promise.resolve({ id });
-      }
-
-      _create (data: Partial<any> | Partial<any>[], _params?: Params) {
-        return Promise.resolve(data);
-      }
-
-      _update (id: NullableId, _data: any, _params?: Params) {
-        return Promise.resolve({ id });
-      }
-
-      _patch (id: NullableId, _data: any, _params?: Params) {
-        return Promise.resolve({ id });
-      }
-
-      _remove (id: NullableId, _params?: Params) {
-        return Promise.resolve({ id });
-      }
-    }
-
     METHODS.forEach(method => {
       it(`${method}`, () => {
         const service = new MethodService({});
@@ -71,77 +24,82 @@ describe('@feathersjs/adapter-commons/service', () => {
       });
     });
 
-    it('does not allow multi patch', () => {
+    it('does not allow multi patch', async () => {
       const service = new MethodService({});
 
-      return service.patch(null, {})
-        .then(() => assert.ok(false))
-        .catch(error => {
-          assert.strictEqual(error.name, 'MethodNotAllowed');
-          assert.strictEqual(error.message, 'Can not patch multiple entries');
-        });
+      await assert.rejects(() => service.patch(null, {}), {
+        name: 'MethodNotAllowed',
+        message: 'Can not patch multiple entries'
+      });
     });
 
-    it('does not allow multi remove', () => {
+    it('does not allow multi remove', async () => {
       const service = new MethodService({});
 
-      return service.remove(null, {})
-        .then(() => assert.ok(false))
-        .catch(error => {
-          assert.strictEqual(error.name, 'MethodNotAllowed');
-          assert.strictEqual(error.message, 'Can not remove multiple entries');
-        });
+      await assert.rejects(() => service.remove(null, {}), {
+        name: 'MethodNotAllowed',
+        message: 'Can not remove multiple entries'
+      });
     });
 
-    it('does not allow multi create', () => {
+    it('does not allow multi create', async () => {
       const service = new MethodService({});
 
-      return service.create([])
-        .then(() => assert.ok(false))
-        .catch(error => {
-          assert.strictEqual(error.name, 'MethodNotAllowed');
-          assert.strictEqual(error.message, 'Can not create multiple entries');
-        });
+      await assert.rejects(() => service.create([], {}), {
+        name: 'MethodNotAllowed',
+        message: 'Can not create multiple entries'
+      });
     });
 
-    it('multi can be set to true', () => {
+    it('multi can be set to true', async () => {
       const service = new MethodService({});
 
       service.options.multi = true;
 
-      return service.create([])
-        .then(() => assert.ok(true));
+      await service.create([]);
     });
   });
 
-  it('filterQuery', () => {
-    const service = new CustomService({
-      whitelist: [ '$something' ]
-    });
-    const filtered = service.filterQuery({
-      query: { $limit: 10, test: 'me' }
-    });
-
-    assert.deepStrictEqual(filtered, {
-      paginate: {},
-      filters: { $limit: 10 },
-      query: { test: 'me' }
+  it('sanitizeQuery', async () => {
+    const service = new MethodService({
+      filters: {
+        $something: true
+      },
+      operators: [ '$test' ]
     });
 
-    const withWhitelisted = service.filterQuery({
-      query: { $limit: 10, $something: 'else' }
+    assert.deepStrictEqual(await service.sanitizeQuery({
+      query: { $limit: '10', test: 'me' } as any
+    }), { $limit: 10, test: 'me' });
+
+    assert.deepStrictEqual(await service.sanitizeQuery({
+      adapter: {
+        paginate: { max: 2 }
+      },
+      query: { $limit: '10', test: 'me' } as any
+    }), { $limit: 2, test: 'me' });
+
+    await assert.rejects(() => service.sanitizeQuery({
+      query: { name: { $bla: 'me' } }
+    }), {
+      message: 'Invalid query parameter $bla'
     });
 
-    assert.deepStrictEqual(withWhitelisted, {
-      paginate: {},
-      filters: { $limit: 10 },
-      query: { $something: 'else' }
-    });
+    assert.deepStrictEqual(await service.sanitizeQuery({
+      adapter: {
+        operators: ['$bla']
+      },
+      query: { name: { $bla: 'Dave' } }
+    }), { name: { $bla: 'Dave' } });
   });
 
   it('getOptions', () => {
-    const service = new AdapterService({
-      multi: true
+    const service = new MethodService({
+      multi: true,
+      paginate: {
+        default: 1,
+        max: 10
+      }
     });
     const opts = service.getOptions({
       adapter: {
@@ -158,20 +116,33 @@ describe('@feathersjs/adapter-commons/service', () => {
       events: [],
       paginate: { default: 10, max: 100 },
       multi: [ 'create' ],
-      filters: [],
-      allow: []
+      filters: {},
+      operators: []
+    });
+
+    const notPaginated = service.getOptions({
+      paginate: false
+    });
+
+    assert.deepStrictEqual(notPaginated, {
+      id: 'id',
+      events: [],
+      paginate: false,
+      multi: true,
+      filters: {},
+      operators: []
     });
   });
 
   it('allowsMulti', () => {
     context('with true', () => {
-      const service = new AdapterService({multi: true});
+      const service = new MethodService({multi: true});
 
-      it('does return true for multible methodes', () => {
+      it('does return true for multiple methodes', () => {
         assert.equal(service.allowsMulti('patch'), true);
       });
 
-      it('does return false for always non-multible methodes', () => {
+      it('does return false for always non-multiple methodes', () => {
         assert.equal(service.allowsMulti('update'), false);
       });
 
@@ -181,13 +152,13 @@ describe('@feathersjs/adapter-commons/service', () => {
     });
 
     context('with false', () => {
-      const service = new AdapterService({multi: false});
+      const service = new MethodService({multi: false});
 
-      it('does return false for multible methodes', () => {
+      it('does return false for multiple methodes', () => {
         assert.equal(service.allowsMulti('remove'), false);
       });
 
-      it('does return true for always multible methodes', () => {
+      it('does return true for always multiple methodes', () => {
         assert.equal(service.allowsMulti('find'), true);
       });
 
@@ -197,17 +168,17 @@ describe('@feathersjs/adapter-commons/service', () => {
     });
 
     context('with array', () => {
-      const service = new AdapterService({multi: ['create', 'get', 'other']});
+      const service = new MethodService({multi: ['create', 'get', 'other']});
 
-      it('does return true for specified multible methodes', () => {
+      it('does return true for specified multiple methodes', () => {
         assert.equal(service.allowsMulti('create'), true);
       });
 
-      it('does return false for non-specified multible methodes', () => {
+      it('does return false for non-specified multiple methodes', () => {
         assert.equal(service.allowsMulti('patch'), false);
       });
 
-      it('does return false for specified always multible methodes', () => {
+      it('does return false for specified always multiple methodes', () => {
         assert.equal(service.allowsMulti('get'), false);
       });