+++ packages/adapter-tests/src/index.ts
@@ -28,7 +28,6 @@ const adapterTests = (testNames: AdapterTestName[]) => {
 
     describe(`Adapter tests for '${serviceName}' service with '${idProp}' id property`, () => {
       after(() => {
-        console.log('\n');
         testNames.forEach(name => {
           if (!allTests.includes(name)) {
             console.error(`WARNING: '${name}' test is not part of the test suite`);