+++ packages/client/test/fixture.ts
@@ -1,6 +1,6 @@
 import { feathers, Application, HookContext, Id, Params } from '@feathersjs/feathers';
 import * as express from '@feathersjs/express';
-import { Service } from '@feathersjs/memory';
+import { MemoryService } from '@feathersjs/memory';
 
 // eslint-disable-next-line no-extend-native
 Object.defineProperty(Error.prototype, 'toJSON', {
@@ -17,7 +17,7 @@ Object.defineProperty(Error.prototype, 'toJSON', {
 });
 
 // Create an in-memory CRUD service for our Todos
-class TodoService extends Service {
+class TodoService extends MemoryService {
   async get (id: Id, params: Params) {
     if (params.query.error) {
       throw new Error('Something went wrong');