+++ packages/schema/test/hooks.test.ts
@@ -5,6 +5,7 @@ describe('@feathersjs/schema/hooks', () => {
   const text = 'Hi there';
 
   let message: MessageResult;
+  let messageOnPaginatedService: MessageResult;
   let user: UserResult;
 
   before(async () => {
@@ -16,6 +17,14 @@ describe('@feathersjs/schema/hooks', () => {
       text,
       userId: user.id
     });
+    messageOnPaginatedService = await app.service('paginatedMessages').create({
+      text,
+      userId: user.id
+    });
+  });
+
+  it('ran resolvers in sequence', async () => {
+    assert.strictEqual(user.name, 'hello (hello@feathersjs.com)');
   });
 
   it('validates data', async () => {
@@ -69,6 +78,64 @@ describe('@feathersjs/schema/hooks', () => {
     });
   });
 
+  it('resolves get result with the object on result', async () => {
+    // eslint-disable-next-line
+    const { password, ...externalUser } = user;
+    const payload = {
+      userId: user.id,
+      text
+    }
+
+    assert.ok(user);
+    assert.strictEqual(user.password, 'hashed', 'Resolved data');
+    assert.deepStrictEqual(message, {
+      id: 0,
+      user,
+      ...payload
+    });
+
+    const result = await app.service('messages').get(0, {
+      provider: 'external'
+    });
+
+    assert.deepStrictEqual(result, {
+      id: 0,
+      user: externalUser,
+      ...payload
+    });
+  });
+
+  it('resolves find results with paginated result object', async () => {
+    // eslint-disable-next-line
+    const { password, ...externalUser } = user;
+    const payload = {
+      userId: user.id,
+      text
+    }
+
+    assert.ok(user);
+    assert.strictEqual(user.password, 'hashed', 'Resolved data');
+    assert.deepStrictEqual(messageOnPaginatedService, {
+      id: 0,
+      user,
+      ...payload
+    });
+
+    const messages = await app.service('paginatedMessages').find({
+      provider: 'external',
+      query: {
+        $limit: 1,
+        $skip: 0
+      }
+    });
+
+    assert.deepStrictEqual(messages, { limit: 1, skip: 0, total: 1, data: [{
+      id: 0,
+      user: externalUser,
+      ...payload
+    }]});
+  });
+
   it('validates and converts the query', async () => {
     const otherUser = await app.service('users').create({
       email: 'helloagain@feathersjs.com',
@@ -81,16 +148,18 @@ describe('@feathersjs/schema/hooks', () => {
     });
 
     const messages = await app.service('messages').find({
+      paginate: false,
       query: {
         userId: `${user.id}`
       }
-    }) as MessageResult[];
+    });
 
     assert.strictEqual(messages.length, 1);
 
     const userMessages = await app.service('messages').find({
+      paginate: false,
       user
-    }) as MessageResult[];
+    });
 
     assert.strictEqual(userMessages.length, 1);
     assert.strictEqual(userMessages[0].userId, user.id);